move date and numeric types to bokeh.core.property.datetime and bokeh.core.property.numeric
